part_1/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
azure-ai-formrecognizer==3.3.3
python-dotenv==1.0.1
pydantic==2.5.0
redis==5.0.1
python-multipart==0.0.6
openai==1.17.0
streamlit==1.32.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
faker==20.1.0